GH-46 Return the response in case of non-101 response from server
Following redirects inside the lib (https://github.com/snapview/tungstenite-rs/pull/148)
has few flows: there is no redirect loop prevention in this case, in case of using the lib in proxy it's impossible to return the upstream response to browser, etc.

With this change response is propagated to the lib's user, so it can decide what to do with it
in case of redirects: either send it to browser for it to follow redirects or implement redirect following on their side.